projects/mnist/pyproject.toml
[project]
name = "mnist"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
